what they said

+++ Pal.Common/ETerritoryType.cs
@@ -5,7 +5,7 @@ namespace Pal.Common
 {
     [SuppressMessage("ReSharper", "UnusedMember.Global")]
     [SuppressMessage("ReSharper", "InconsistentNaming")]
-    public enum ETerritoryType : ushort
+    public enum ETerritoryType : uint
     {
         Palace_1_10 = 561,
         Palace_11_20,
